Factor the equality tests into their own suite.
