+++ BoxTests/BoxTests.swift
@@ -8,16 +8,4 @@ class BoxTests: XCTestCase {
 		let box = Box(1)
 		XCTAssertEqual(box.value, 1)
 	}
-
-	func testEquality() {
-		let (a, b, c) = (Box(1), Box(1), Box(2))
-		XCTAssertTrue(a == b)
-		XCTAssertFalse(b == c)
-	}
-
-	func testInequality() {
-		let (a, b, c) = (Box(1), Box(1), Box(2))
-		XCTAssertFalse(a != b)
-		XCTAssertTrue(b != c)
-	}
 }
