+++ BoxTests/BoxTypeTests.swift
@@ -1,0 +1,18 @@
+//  Copyright (c) 2014 Rob Rix. All rights reserved.
+
+import Box
+import XCTest
+
+class BoxTypeTests: XCTestCase {
+	func testEquality() {
+		let (a, b, c) = (Box(1), Box(1), Box(2))
+		XCTAssertTrue(a == b)
+		XCTAssertFalse(b == c)
+	}
+
+	func testInequality() {
+		let (a, b, c) = (Box(1), Box(1), Box(2))
+		XCTAssertFalse(a != b)
+		XCTAssertTrue(b != c)
+	}
+}
